Created loader abstraction for loadOne
